Rename network_graph.rs to graph.rs
