+++ src/main/network/mod.rs
@@ -1,4 +1,4 @@
-pub mod network_graph;
+pub mod graph;
 mod packet;
 mod petgraph_wrapper;
 mod relay;
